@@ -1,7 +1,6 @@
 package providers
 
 import (
-	"context"
 	"slices"
 	"testing"
 
@@ -63,7 +62,7 @@ funckyfunc := true
 		},
 	}
 
-	completions, err := p.Run(context.Background(), c, completionParams, nil)
+	completions, err := p.Run(t.Context(), c, completionParams, nil)
 	if err != nil {
 		t.Fatalf("Unexpected error: %v", err)
 	}
