@@ -63,7 +63,7 @@ rules:
 	}
 
 	// set up the server and client connections
-	ctx, cancel := context.WithCancel(context.Background())
+	ctx, cancel := context.WithCancel(t.Context())
 	defer cancel()
 
 	_, connClient := createAndInitServer(t, ctx, newTestLogger(t), tempDir, clientHandler)
